refactor: BowlingGame 제거 & Pin 추가
