--- src/main/java/bowling/domain/BowlingGame.java
@@ -1,5 +1,0 @@
-package bowling.domain;
-
-public class BowlingGame {
-    public static final int PIN_COUNT = 10;
-}
