+++ src/main/java/bowling/domain/pin/Pin.java
@@ -1,0 +1,5 @@
+package bowling.domain.pin;
+
+public class Pin {
+    public static final int PIN_COUNT = 10;
+}
